requirements.txt
certifi==2022.12.7
babel==2.11.0
flask-babel==3.0.1
flask==2.2.3
jinja2==3.1.2
lxml==4.9.2
pygments==2.14.0
python-dateutil==2.8.2
pyyaml==6.0
httpx[http2]==0.21.2
Brotli==1.0.9
uvloop==0.17.0
httpx-socks[asyncio]==0.7.2
setproctitle==1.3.2
redis==4.5.1
markdown-it-py==2.1.0
typing_extensions==4.5.0
fasttext-predict==0.9.2.1
